+++ packages/koot/utils/get-dist-path.js
@@ -1,7 +1,8 @@
-const path = require('path')
-const getCwd = require('./get-cwd')
+const fs = require('fs-extra');
+const path = require('path');
+const getCwd = require('./get-cwd');
 
-let p
+let p;
 
 /**
  * 获取打包结果路径
@@ -10,9 +11,21 @@ let p
 module.exports = () => {
     // console.log('global.KOOT_DIST_DIR', global.KOOT_DIST_DIR)
     if (typeof p !== 'string') {
-        p = typeof global.KOOT_DIST_DIR === 'string'
-            ? global.KOOT_DIST_DIR
-            : path.resolve(getCwd(), process.env.KOOT_DIST_DIR)
+        p =
+            typeof global.KOOT_DIST_DIR === 'string'
+                ? global.KOOT_DIST_DIR
+                : (() => {
+                      let cwd = getCwd();
+                      let parent = path.resolve(cwd, '..');
+                      let result = path.resolve(cwd, process.env.KOOT_DIST_DIR);
+                      while (!fs.existsSync(result) && cwd !== parent) {
+                          cwd = parent;
+                          parent = path.resolve(cwd, '..');
+                          result = path.resolve(cwd, process.env.KOOT_DIST_DIR);
+                      }
+                      if (fs.existsSync(result)) return result;
+                      return path.resolve(cwd, process.env.KOOT_DIST_DIR);
+                  })();
     }
-    return p
-}
+    return p;
+};