+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 -   **新特性**
     -   现支持使用 _TypeScript_ 开发 _React_ 组件。详情请参见文档 [TypeScript 开发](https://koot.js.org/#/typescript)
     -   `extend()` React 高阶组件新增选项 `ssr`：可控制对应组件的 SSR 行为。详情请参见文档 [React 开发](https://koot.js.org/#/react?id=参数)
+    -   SPA
+        -   现在生产环境下会创建一个简易服务器的启动脚本文件（位于打包目录下的 `/.server/index.js`）
     -   **新配置项** `sessionStore` - 将全部或部分 _state_ 对象暂存在 `sessionStorage` 中，在刷新页面后这些 _state_ 会自动还原。详情请参见文档 [配置/sessionStore](https://koot.js.org/#/config?id=sessionStore)
     -   **新函数** `createStore()` - 方便项目更便捷的创建 _Redux store_
         -   `import { createStore } from 'koot';`
@@ -38,9 +40,12 @@
         -   `url-loader` -> _2.0.1_
         -   `os-locale` -> _4.0.0_
     -   minor
+        -   `cli-spinners` -> _2.2.0_
+        -   `fs-extra` -> _8.1.0_
         -   `inquirer` -> _6.4.1_
         -   `mini-css-extract-plugin` -> _0.7.0_
-        -   `react-hot-loader` -> _4.11.1_
+        -   `open` -> _6.4.0_
+        -   `react-hot-loader` -> _4.12.0_
         -   `webpack` -> _4.35.0_
         -   `webpack-dev-server` -> _3.7.2_
     -   patch
@@ -54,6 +59,8 @@
 -   SSR
     -   现在服务器环境的打包结果中，不再会出现静态资源文件
 -   更新依赖包
+    -   major
+        -   `svg-url-loader` -> _3.0.0_
     -   minor
         -   `eslint-config-koot` -> _0.2.0_
 
@@ -61,6 +68,7 @@
 
 -   更新依赖包
     -   minor
+        -   `fs-extra` -> _8.1.0_
         -   `inquirer` -> _6.4.1_
     -   patch
         -   `semver` -> _6.1.2_